feat: Refactor code and update language file
