+++ src/app/questionnaire/page.tsx
@@ -1,35 +1,15 @@
-"use client";
-import { useEffect } from "react";
-import { ControlState, useControl } from "@/store/control";
-import isPhone from "@/utils/isPhone";
-import { t } from "i18next";
+"use server";
+import { Box, Table } from "@radix-ui/themes";
 
 const QuestionnairePage = () => {
-    const changeControl = useControl(state => (state as ControlState).changeControl);
-    const clear = useControl(state => (state as ControlState).clear);
-    const changeShow = useControl(state => (state as ControlState).changeShow);
-
-    useEffect(() => {
-        changeControl([
-            {
-                type: "group",
-                name: t("Guide"),
-            },
-            {
-                type: "group",
-                name: t("Questionnaire"),
-            },
-        ]);
-        return () => {
-            clear();
-            if (!isPhone()) {
-                // isn't phone
-                changeShow(true);
-            }
-        };
-    }, [changeControl, changeShow, clear]);
-
-    return <div>111</div>;
+    return (
+        <main className="w-full h-full p-5 sm:p-10">
+            <Box className="">
+                {/* <Table.Root></Table.Root> */}
+                {/* pass in next */}
+            </Box>
+        </main>
+    );
 };
 
 export default QuestionnairePage;
