+++ src/language/en-us.json
@@ -5,6 +5,5 @@
     "Home": "Home",
     "About": "About",
     "Getting Started": "Getting Started",
-    "Guide": "Guide",
-    "Questionnaire": "Questionnaire"
+    "My": "My"
 }
