+++ src/language/zh-cn.json
@@ -5,6 +5,5 @@
     "Home": "首页",
     "About": "关于",
     "Getting Started": "快速开始",
-    "Guide": "概况",
-    "Questionnaire": "问卷"
+    "My": "我的"
 }
